feat: radar skeleton
